remove: header.js로 통합
