+++ src/layouts/header/header.scss
@@ -5,7 +5,7 @@ body {
   height: auto;
   position: relative;
   color: $gray-200;
-  background-color: black;
+  // background-color: black;
 }
 
 .header {
