+++ src/layouts/header/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>header</title>
     <script src="https://kit.fontawesome.com/72b5af3d51.js" crossorigin="anonymous"></script>
-    <script src="./headerComponent.js" type="module"></script>
     <script src="./header.js" type="module"></script>
 </head>
 <body>
